feat: 네비게이션바 프로토콜화
