+++ Baton/Source/Presentation/Common/Navigation/BatonNavigationBackButton.swift
@@ -1,0 +1,14 @@
+import UIKit
+
+struct BatonNavigationButton {
+    static func backButton(target: Any, action: Selector) -> UIBarButtonItem {
+        let button = UIBarButtonItem(
+            image: UIImage(systemName: "chevron.left"),
+            style: .plain,
+            target: target,
+            action: action
+        )
+        button.tintColor = .bblack
+        return button
+    }
+}
